# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestionar kits de sensores y sus datos.",
        "title": "API Hexagonal Go (Sensor Kits)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.tu-soporte.com/support",
            "email": "support@tu-dominio.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080  // Cambia esto por tu host y puerto reales (puedes usar variables de entorno)",
    "basePath": "/v1             // Base path de tu API v1",
    "paths": {
...
        "/data/temperature/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new temperature level data record for a specific kit.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Temperature Data"
                ],
                "summary": "Register temperature data record",
                "parameters": [
                    {
                        "description": "temperature data record",
                        "name": "temperatureData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RegisterTemperatureRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "temperature record registered successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entities.TemperatureData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body, validation failed, or invalid Kit ID",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while registering temperature record",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/data/temperature/kit/{kit_id}/minutes/{minutes}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves temperature data records for a specific kit within the last N minutes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Temperature Data"
                ],
                "summary": "Get recent temperature data records",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Kit ID",
                        "name": "kit_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Number of minutes ago to fetch records from",
                        "name": "minutes",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "temperature records retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entities.TemperatureData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Kit ID or minutes parameter provided",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (token missing or invalid)",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error while retrieving temperature records",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.TemperatureData": {
            "type": "object",
            "properties": {
                "data_id": {
                    "description": "Corresponds to data_id PK",
                    "type": "integer"
                },
                "humidity": {
                    "description": "Humidity reading",
                    "type": "number"
                },
                "kit_id": {
                    "description": "Foreign key to kits table",
                    "type": "integer"
                },
                "temperature": {
                    "description": "Temperature reading",
                    "type": "number"
                },
                "timestamp": {
                    "description": "Timestamp from DB default",
                    "type": "string"
                }
            }
        },
...
        "request.RegisterTemperatureRequest": {
            "type": "object",
            "required": [
                "humidity",
                "kit_id",
                "temperature"
            ],
            "properties": {
                "humidity": {
                    "description": "Use appropriate validation (e.g., numeric range 0-100) if needed",
                    "type": "number"
                },
                "kit_id": {
                    "description": "Kit ID must be provided and positive",
                    "type": "integer"
                },
                "temperature": {
                    "description": "Use appropriate validation (e.g., numeric range) if needed",
                    "type": "number"
                }
            }
        },
        "responses.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}